Move home api onto migrations tooling
