+++ home_api/BUILD.bazel
@@ -1,0 +1,10 @@
+load("@rules_python//python:defs.bzl", "py_library")
+
+py_library(
+    name = "config_py",
+    srcs = ["config.py"],
+    visibility = ["//home_api:__subpackages__"],
+    deps = [
+        "//base:flask_app_py",
+    ],
+)
